fix : user entity persistence
