--- src/main/java/com/flab/jbly/infra/jpa/UserJpaRepository.java
@@ -1,10 +1,0 @@
-package com.flab.jbly.infra.jpa;
-
-import com.flab.jbly.domain.user.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserJpaRepository extends JpaRepository<User, Long> {
-
-    boolean existsByUserId(String userId);
-
-}
