+++ src/main/java/com/flab/jbly/infrastructure/jpa/user/UserRepositoryAdapter.java
@@ -9,10 +9,20 @@ import org.springframework.stereotype.Repository;
 @RequiredArgsConstructor
 public class UserRepositoryAdapter implements UserRepository {
 
-		private final UserJpaRepository repository;
+    private final UserJpaRepository repository;
 
-		@Override
-		public User findByUserId(String userId) {
-				return repository.findByUserId(userId);
-		}
+    @Override
+    public User findByUserId(String userId) {
+        return repository.findByUserId(userId);
+    }
+
+    @Override
+    public User save(User user) {
+        return repository.save(user);
+    }
+
+    @Override
+    public boolean existsByUserId(String userId) {
+        return repository.existsByUserId(userId);
+    }
 }
